requirements.txt
flask
Flask
opencv-python
tensorflow
numpy
pymysql
flask-cors
Flask-Session
pillow
python-dotenv
tensorflow_hub
tensorflow_hub
rembg
opencv-python
numpy==1.23.5
numba
flask-jwt-extended